refactor: simplify state with zustand
